# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API untuk sistem pelaporan keuangan dengan manajemen users, transaksi, dan master data",
        "title": "Pelaporan Keuangan API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/status-transaksi": {
            "post": {
                "description": "Create a new status transaksi with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-StatusTransaksi"
                ],
                "summary": "Create new status transaksi",
                "parameters": [
                    {
                        "description": "Status transaksi input data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InputStatusTransaksi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success create status transaksi",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or validation error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            }
        },
        "/status-transaksi/{id}": {
            "get": {
                "description": "Retrieve detailed information of a specific status transaksi by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-StatusTransaksi"
                ],
                "summary": "Get status transaksi by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Status Transaksi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success get status transaksi detail",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/helpers.ResponseGetDetailSuccess"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/dtos.ResStatusTransaksi"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid ID parameter",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Status transaksi not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing status transaksi with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-StatusTransaksi"
                ],
                "summary": "Update existing status transaksi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Status Transaksi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Status transaksi update data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InputStatusTransaksi"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success update status transaksi",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid request data or validation error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Status transaksi not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete an existing status transaksi by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MasterData-StatusTransaksi"
                ],
                "summary": "Delete status transaksi",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Status Transaksi ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success delete status transaksi",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseCUDSuccess"
                        }
                    },
                    "400": {
                        "description": "Invalid ID parameter",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "404": {
                        "description": "Status transaksi not found",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/helpers.ResponseError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dtos.InputStatusTransaksi": {
            "type": "object",
            "required": [
                "nama"
            ],
            "properties": {
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "dtos.ResStatusTransaksi": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "helpers.ResponseCUDSuccess": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "helpers.ResponseError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "error message"
                },
                "status": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "helpers.ResponseGetDetailSuccess": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "success message"
                },
                "status": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}